requirements.txt
pyfcm
pywebpush
numpy
flask
flask_login
pyyaml
pandas
werkzeug
sqlalchemy
requests
flask_sqlalchemy
flask_wtf
wtforms
email_validator
psycopg2-binary
flask_cachebuster